+++ src/components/Keyword/KeywordFilter/styles.js
@@ -124,29 +124,6 @@ export const FilterItemImage = styled.img`
     margin-right:5px;
 `;
 
-export const SearchInput = styled.input`
-    width:665px;
-    padding:8px;
-    border:none;
-    background-color:#eeeeee;
-`;
-
-export const SearchButton = styled.button`
-    width:101px;
-    height:34px;
-    padding:8px;
-    display:flex;
-    justify-content:center;
-    align-items:center;
-    font-size:12px;
-    background-color:#222222;
-    color:#ffffff;
-`;
-
-export const SearchImage = styled.img`
-    margin-left:8px;
-`;
-
 export const MainList = styled.div`
     position:absolute;
     left:${props=>props.toggle?'501px':'366px'};
