+++ src/components/Keyword/KeywordSearch/styles.js
@@ -1,0 +1,24 @@
+import styled from "styled-components";
+
+export const SearchInput = styled.input`
+    width:665px;
+    padding:8px;
+    border:none;
+    background-color:#eeeeee;
+`;
+
+export const SearchButton = styled.button`
+    width:101px;
+    height:34px;
+    padding:8px;
+    display:flex;
+    justify-content:center;
+    align-items:center;
+    font-size:12px;
+    background-color:#222222;
+    color:#ffffff;
+`;
+
+export const SearchImage = styled.img`
+    margin-left:8px;
+`;
